refactor(cart): /domain/cart/model/dto dto 수정
class -> record 수정, 필요 없는 DTO 삭제

ref#9
