--- src/main/java/app/domain/cart/model/dto/request/UpdateCartItemRequest.java
@@ -1,8 +1,0 @@
-package app.domain.cart.model.dto.request;
-
-import lombok.Data;
-
-@Data
-public class UpdateCartItemRequest {
-	private int quantity;
-}
