+++ src/main/java/app/domain/cart/service/impl/CartServiceImpl.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import app.domain.cart.model.CartItemRepository;
 import app.domain.cart.model.CartRepository;
-import app.domain.cart.model.dto.RedisCartItem;
+import app.domain.cart.model.dto.response.RedisCartItem;
 import app.domain.cart.model.entity.Cart;
 import app.domain.cart.model.entity.CartItem;
 import app.domain.cart.service.CartRedisService;
